Move test module to lapin_test.exs
